fixes code review
